Refactor to split concerns of commands, models and serialization
